LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-badInput-expected.txt


PASS [INPUT in EMAIL status] The multiple attribute is false and the value attribute is empty 
PASS [INPUT in EMAIL status] The multiple attribute is false and the value attribute is a valid e-mail address 
PASS [INPUT in EMAIL status] The multiple attribute is true and the value contains valid e-mail addresses 
PASS [INPUT in EMAIL status] The multiple attribute is true and the value attribute contains a ',' 
PASS [INPUT in DATETIME-LOCAL status] The value attribute is empty 
PASS [INPUT in DATETIME-LOCAL status] The value attribute is a valid date and time string 
PASS [INPUT in DATETIME-LOCAL status] The value attribute cannot convert to a valid normalized forced-UTC global date and time string 
PASS [INPUT in COLOR status] The value attribute is empty 
PASS [INPUT in COLOR status] The value attribute is a valid sample color 
PASS [INPUT in COLOR status] The value attribute is not a valid lowercase sample color 
PASS [INPUT in COLOR status] The value attribute cannot convert to a valid sample color 
